requirements.txt
discord.py
forecastiopy